# openapi.yaml
openapi: 3.0.3
info:
  title: MultiSig Treasury API
  version: 1.0.0
  description: |
    REST API for managing multi-signature treasuries on Sui blockchain.
    Supports programmable spending policies, time-locked proposals, and emergency procedures.
  contact:
    name: API Support
    email: support@example.org
servers:
  - url: https://api.example.org/v1
    description: Production server
  - url: http://localhost:3000/v1
    description: Development server
...
paths:
  /treasury/create:
    post:
      summary: Create a new treasury
      description: Initialize a new multi-signature treasury with signers and policies
      tags:
        - Treasury
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTreasuryRequest'
            examples:
              basic:
                summary: Basic treasury
                value:
                  signers: ["0x1234", "0x5678", "0x9abc"]
                  threshold: 2
              with_emergency:
                summary: Treasury with emergency signers
                value:
                  signers: ["0x1234", "0x5678", "0x9abc"]
                  threshold: 2
                  emergency_signers: ["0x1234", "0x5678"]
                  emergency_threshold: 2
      responses:
        '201':
          description: Treasury created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Treasury'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /treasury/{id}:
    get:
      summary: Get treasury details
      tags:
        - Treasury
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Treasury ID (Sui object ID)
      responses:
        '200':
          description: Treasury details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Treasury'
        '404':
          description: Treasury not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /treasury/{id}/deposit:
    post:
      summary: Deposit funds into treasury
      tags:
        - Treasury
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositRequest'
      responses:
        '200':
          description: Deposit successful
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /treasury/{id}/balance:
    get:
      summary: Get treasury balances
      tags:
        - Treasury
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Treasury balances
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
  /treasury/{id}/signers:
    put:
      summary: Update treasury signers
      description: Add or remove signers (requires multi-sig approval)
      tags:
        - Treasury
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSignersRequest'
      responses:
        '200':
          description: Signers updated successfully
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    CreateTreasuryRequest:
      type: object
      properties:
        signers:
          type: array
          items:
            type: string
          minItems: 2
          description: List of signer addresses
        threshold:
          type: integer
          minimum: 2
          description: Number of signatures required
        emergency_signers:
          type: array
          items:
            type: string
          description: Optional emergency signers
        emergency_threshold:
          type: integer
          description: Emergency signature threshold
      required:
        - signers
        - threshold
    Treasury:
      type: object
      properties:
        id:
          type: string
          description: Treasury object ID
        signers:
          type: array
          items:
            type: string
        threshold:
          type: integer
        emergency_signers:
          type: array
          items:
            type: string
        emergency_threshold:
          type: integer
        balances:
          type: array
          items:
            $ref: '#/components/schemas/Balance'
        policies:
          type: array
          items:
            type: string
          description: Policy IDs
        created_at:
          type: integer
          format: int64
    Balance:
      type: object
      properties:
        coin_type:
          type: string
        amount:
          type: string
          description: Amount in smallest unit
    DepositRequest:
      type: object
      properties:
        coin_id:
          type: string
          description: Coin object ID to deposit
        amount:
          type: string
      required:
        - coin_id
        - amount
    BalanceResponse:
      type: object
      properties:
        treasury_id:
          type: string
        balances:
          type: array
          items:
            $ref: '#/components/schemas/Balance'
    UpdateSignersRequest:
      type: object
      properties:
        add_signers:
          type: array
          items:
            type: string
        remove_signers:
          type: array
          items:
            type: string
        new_threshold:
          type: integer
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string